Break out implicit synthesis into separate class
